// web/frontend/src/app/shared/models/scene-status.model.ts
export enum SceneStatus {
  CREATED = 'CREATED',
  VIDEO_UPLOADED = 'VIDEO_UPLOADED',
  BBOX_SELECTED = 'BBOX_SELECTED',
  SEGMENTING = 'SEGMENTING',
  MASKS_READY = 'MASKS_READY',
  DATASET_READY = 'DATASET_READY',
  COLMAP_RUNNING = 'COLMAP_RUNNING',
  COLMAP_READY = 'COLMAP_READY',
  TRAINING_3DGS = 'TRAINING_3DGS',
  THREE_DGS_READY = 'THREE_DGS_READY',
  TRAINING_SUGAR = 'TRAINING_SUGAR',
  SUGAR_READY = 'SUGAR_READY',
  COMPLETED = 'COMPLETED',
  ERROR = 'ERROR',
  CANCELED = 'CANCELED'
}